Groups (#76)
* Remove compile warnings

* v0.8.0

* Update client version to 0.8

* Add list of available groups (front-end)

Displaying the capacity, title & author.

Backend is not implemented.

* Backend for groups (changed name from Sessions)

Placeholders for controllers.

* Add Group when logged in

Update group icon.

* Move click handler from icon to the button

Closes #66

* Expand groups API

* List "my" groups above available with ability to manage

* Add front-end for creating a new group

* Comment out group visibility for now

* Fix dependancies

The "react-dom-router" was in the top level package.json instead of client/ directory.

* Migrate to new Material-UI

* Groups backend

* Groups list front-end

Closes #70.

Alter SavedDatabase component to handle errors better.

* Clean-up

* Ability to join a group

Remove active groups from the "all group" list.

Prevent group owner from leaving there group (have to delete).

Closes #43.

* Add user context to Schema and LoggedIn component

In preperation for #71.

* Distinction in Groups component for "current" group

Ability to leave the current group.

Current groups gets a star in the list.

Closes #74, and closes #73.

* Seperate SavedDatabase component from the SavedDatabaseList

* Split SavedDatabase into different file to help readabilit

Clean up

* Progress on creating a group

Allows choosing of a SavedDatabase

* Bugfix: Return a resolved promise instead of setState

setState returns "undefined".

* Ability to create a group

Added the submit handler to post the request to the server.

* Inject UserContext into Main component to access inside lifecycle methods

* Bugfix find duplicate user group session

* Stop closing the Group dialog on joining & leaving of a group

* Add ability to get group information

* Ability to save question progress while in a group (1/2)

* Progress on #72

* Fix join group

* Push built question to server if none already stored

Use existing questions if the user has existing process.

Add new routes & controller actions to update a group, and remove a user.

Add middleware to check if the user can manage a group.

* Fix sending of group progress to server

* Send all users back to the server when retrieving a group

* Alter the data when fetching a group

* List all group users and display their progress

Ability to remove a user from a group.

Ability to update a group title.

Closes #44.

* Correctly destroy session on logout

* Remove compile warnings

* v0.8.0

* Update client version to 0.8

+++ controllers/database.js
@@ -7,6 +7,7 @@ const config = require("../config/config");
 
 // Models
 const Database = require("../models/Database");
+const Group = require("../models/Group");
 
 const { check, validationResult } = require("express-validator/check");
 
@@ -35,9 +36,7 @@ exports.canSaveDatabase = (req, res, next) => {
     // Check if the user has reached their upload limit
     if (databases.length >= config.database.limit) {
       return res.status(403).json({
-        errors: {
-          length: { msg: "You have reached the limit of saved databases" }
-        }
+        error: { message: "You have reached the limit of saved databases" }
       });
     }
 
@@ -50,14 +49,11 @@ exports.saveDatabase = (req, res, next) => {
   const database = new Database({
     title: req.params.title,
     path: req.file.filename,
-
     creator: req.user.id
   });
 
   database.save(err => {
-    if (err) {
-      return next(err);
-    }
+    if (err) return next(err);
   });
 
   return res.json(database);
@@ -67,6 +63,12 @@ exports.loadDatabase = (req, res, next) => {
   const { id } = req.params;
 
   Database.findById(id, (err, database) => {
+    if (err) return next(err);
+
+    if (!database) {
+      return res.sendStatus(404);
+    }
+
     const filename = database.path;
 
     const options = {
@@ -87,15 +89,27 @@ exports.loadDatabase = (req, res, next) => {
 exports.deleteDatabase = (req, res, next) => {
   const { id } = req.params;
 
-  Database.findOneAndRemove(
-    { _id: id, creator: req.user.id },
-    (err, database) => {
-      if (err) return next(err);
+  Group.findOne({ database: id }, (err, dependantGroup) => {
+    if (err) return next(err);
 
-      // Remove the file from the server too.
-      fs.unlink(`./saves/${database.path}`);
-
-      return res.sendStatus(200);
+    if (dependantGroup) {
+      return res.status(400).json({
+        error: {
+          message: "Cannot delete a database that a group depends upon."
+        }
+      });
     }
-  );
+
+    Database.findOneAndRemove(
+      { _id: id, creator: req.user.id },
+      (err, database) => {
+        if (err) return next(err);
+
+        // Remove the file from the server too.
+        fs.unlink(`./saves/${database.path}`);
+
+        return res.sendStatus(200);
+      }
+    );
+  });
 };
